requirements.txt
Django==2.2
django-tenant-schemas==1.9.0
psycopg2==2.8.1
psycopg2-binary==2.8.1
pytz==2018.9
sqlparse==0.3.0
python-tbk==0.2.1
